Update GoCardlessV2RedirectPaymentDriver.php
This fixes an issue were GoCardless payments were not able to auto collect with recurring invoices.  This fix only applies to new mandates created on GoCardless.  To fix already existing mandates you will need to update payment_methods.payment_type_id in the database, and change it from 5 to 31 ONLY when it's a GoCardless record.